requirements/requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
Django==1.11.29
djangorestframework==3.9.4
enum34==1.1.6
gyp==0.1
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
mysql-connector-python==2.1.6
mysql-utilities==1.6.4
paramiko==2.0.0
pexpect==4.2.1
pyasn1==0.4.2
pycrypto==2.6.1
pygobject==3.26.1
pyodbc==4.0.17
pysqlite==2.7.0
pytz==2020.5
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
